+++ IBAnimatableApp/IBAnimatableApp/Playground/UserInterfaces/Gradients/GradientViewController.swift
@@ -14,12 +14,14 @@ final class GradientViewController: UIViewController {
   let startPointValues = ParamType.enumeration(values: ["top", "topLeft", "topRight", "left", "right", "bottom", "bottomRight", "bottomLeft"])
   let colorValues = ParamType(fromEnum: ColorType.self)
   var usePredefinedGradient = true
+  var useRadialGradient = false
   lazy var componentValues: [ParamType] = {
     self.usePredefinedGradient ? [self.gradientValues, self.startPointValues] : [self.colorValues, self.colorValues, self.startPointValues]
   }()
 
   override func viewDidLoad() {
     super.viewDidLoad()
+    gView.gradientMode = useRadialGradient ? .radial : .linear
     if usePredefinedGradient {
       gView.predefinedGradient = GradientType(rawValue: gradientValues.value(at: 0))
       gView.startPoint = GradientStartPoint(string: startPointValues.value(at: 0))
@@ -41,7 +43,7 @@ final class GradientViewController: UIViewController {
 
 }
 
-extension GradientViewController : UIPickerViewDelegate, UIPickerViewDataSource {
+extension GradientViewController: UIPickerViewDelegate, UIPickerViewDataSource {
   func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
     return componentValues[component].count()
   }