+++ Sources/Extensions/UIColorExtension.swift
@@ -12,7 +12,7 @@ public extension UIColor {
     var int = UInt32()
     Scanner(string: hex).scanHexInt32(&int)
     let a, r, g, b: UInt32
-    switch hex.characters.count {
+    switch hex.count {
     case 3:
       (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
     case 6: