+++ Sources/Views/AnimatableScrollView.swift
@@ -157,6 +157,12 @@ open class AnimatableScrollView: UIScrollView, CornerDesignable, FillDesignable,
   @IBInspectable open var toneOpacity: CGFloat = CGFloat.nan
 
   // MARK: - GradientDesignable
+  open var gradientMode: GradientMode = .linear
+  @IBInspectable var _gradientMode: String? {
+    didSet {
+      gradientMode = GradientMode(string: _gradientMode) ?? .linear
+    }
+  }
   @IBInspectable open var startColor: UIColor?
   @IBInspectable open var endColor: UIColor?
   open var predefinedGradient: GradientType?