+++ Sources/Views/AnimatableTableView.swift
@@ -71,6 +71,12 @@ open class AnimatableTableView: UITableView, FillDesignable, BorderDesignable, G
     }
   }
   // MARK: - GradientDesignable
+  open var gradientMode: GradientMode = .linear
+  @IBInspectable var _gradientMode: String? {
+    didSet {
+      gradientMode = GradientMode(string: _gradientMode) ?? .linear
+    }
+  }
   @IBInspectable open var startColor: UIColor?
   @IBInspectable open var endColor: UIColor?
   open var predefinedGradient: GradientType?